[WIP] Working on selection menu for the order of tasks has been added.
